+++ src/v1/errors/PackageErrorTable.ts
@@ -31,12 +31,7 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import {
-    ErrorTable,
-    ErrorTableTemplateWithNoExtraData,
-    ExtraDataTemplate,
-    NoExtraDataTemplate,
-} from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
+import { ErrorTable, ErrorTableTemplate } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 import { httpStatusCodeFrom } from "@ganbarodigital/ts-lib-http-types/lib/v1";
 import { packageNameFrom } from "@ganbarodigital/ts-lib-packagename/lib/v1";
 
@@ -44,19 +39,15 @@ import { InvalidUuidTemplate } from "./InvalidUuid";
 
 const PACKAGE_NAME = packageNameFrom("@ganbarodigital/ts-uuid-parser/lib/v1");
 
+type PackageErrorTableIndex<T extends ErrorTable> = ErrorTableTemplate<T, string>;
 export class PackageErrorTable implements ErrorTable {
-    [key: string]: ErrorTableTemplateWithNoExtraData<ErrorTable, string, ExtraDataTemplate | NoExtraDataTemplate>;
+    [key: string]: PackageErrorTableIndex<PackageErrorTable>;
 
     public "invalid-uuid": InvalidUuidTemplate = {
         packageName: PACKAGE_NAME,
         errorName: "invalid-uuid",
         detail: "UUID is invalid / not in RFC 4122 format",
         status: httpStatusCodeFrom(422),
-        extra: {
-            public: {
-                invalidInput: "the invalid input goes here",
-            },
-        },
     };
 }